+++ src/Routes.js
@@ -12,6 +12,7 @@ import IndividualEvent from './views/pages/events/individualEvent';
 import Register from './views/pages/register/Register';
 import CreateNewEvent from './components/CreateNewEvent';
 import Settings from './views/pages/settings/Settings';
+import LandingPage from './views/pages/HomeView/LandingPage';
 
 const renderRoutes = () => (
   <Suspense fallback={<LoadingScreen />}>
@@ -29,6 +30,7 @@ const renderRoutes = () => (
       <Route path="/dashboard" exact render={() => <Navigation />} />
       <Route path="/profile" exact render={() => <Profile />} />
       <Route path="/events" exact render={() => <EventDefaultPage />} />
+      <Route path="/landing" exact render={() => <LandingPage />} />
       <Route
         path="/events/individual-event"
         exact
