+++ src/components/Logo.js
@@ -1,9 +1,17 @@
 import React from 'react';
-
-function Logo(props) {
-  return (
-    <img alt="Logo" height="32px" src="/static/logo/logo.svg" {...props} />
-  );
+import PropTypes from 'prop-types';
+function Logo(props, variant) {
+  if (variant === "black") {
+    return (
+      <img alt="Logo" height="32px" src="/static/logo/logo.png" {...props} />
+    );
+  }
+  else {
+    return(
+    <img alt="Logo" height="32px" src="/static/logo/logo.svg" {...props} />);
+  }
 }
-
+Logo.propTypes = {
+  variant: PropTypes.string,
+};
 export default Logo;
